requirements.txt
aws-cdk.core==1.128.0
aws-cdk.aws_ec2==1.128.0
aws-cdk.aws_events==1.128.0
aws-cdk.aws_events_targets==1.128.0
aws-cdk.aws_iam==1.128.0
aws-cdk.aws_lambda==1.128.0
aws-cdk.aws_s3==1.128.0
aws-cdk.aws_sns==1.128.0
aws-cdk.aws_sns_subscriptions==1.128.0
aws-cdk.aws_stepfunctions==1.128.0
aws-cdk.aws_stepfunctions_tasks==1.128.0
pyyaml<5.5, >=3.10
boto3==1.19.6
botocore==1.22.12